Added model data to model window

+++ 3DViewerX/Viewport.h
@@ -20,8 +20,8 @@ private:
 	int m_Width = 800;
 	int m_Height = 600;
 
-	int m_WindowPosX = 0;
-	int m_WindowPosY = 0;
+	float m_WindowPosX = 0;
+	float m_WindowPosY = 0;
 
 	bool m_IsViewportFocused = false;
 };
